chore: Add the thing ID input back
The Thing ID input was replaced with an hardcoded variable to allow for
a single variable to change all panels in a dashboard. This however
broke the Alerts & Explore pages as there is no way to specify variables
there.

This commit reintroduces the Thing ID input and adds supports for
templating to both inputs, allowing the input to be configured to use a
variable value.